Start writing a completion flag after successful download
This will prevent us from making a bunch of requests to Vimeo for
videos we've already successfully downloaded, which triggers some kind
of rate limiter on their side that manifests as 404s.

With this, the script will only proceed as quickly as it can download
the actual files, which is slow enough to run indefinitely without
hitting any rate limits.